refactor `DependencyResolutionProcess` to recursively resolve types
Instead of doing specific actions like iterating all annotations or traversing up the class hierarchy we can just control the type registering process and execute runs for a specific number of times. E.g. we registered all member types on the initial import, so when we by default execute only 1 run of registering member types that means that the types resolved through the member path won't have their own members resolved in turn. We can mimic the current behavior by configuring the different `maxRuns` parameters to do 1 run for member types and access types (i.e. break the resolution after a chain length of 1) and infinitely many runs for inheritance and annotation runs. This way we will traverse inheritance of all types all the way and annotations including their parameters all the way. One benefit is that this opens the way to make this configurable. I.e. users can then control in a simple way how far they want to traverse the graph and decide about the tradeoff with regard to performance themselves.

The way this whole process works is that further resolution will in turn invoke the import with the original `DeclarationHandler`. So, when we start the next run we check `classes.ensureIfPresent(..)`, if this fails the class is resolved with the configured `ClassResolver` (by default from classpath) which in turn will again register additional type names to resolve. So in each run we can configure which such type names we are still interested in or where we want to break the chain.

Note, that this also fixes the problem that annotation class parameters were not imported correctly. The recursive nature of the annotation parameter resolution now ensures that all `Class<?>` parameters of annotations are automatically resolved.

Signed-off-by: Peter Gafert <peter.gafert@tngtech.com>